requirements/requirements_zeta_ia.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alabaster==1.0.0
annotated-doc==0.0.4
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astroid==4.0.4
asttokens==3.0.1
async-lru==2.3.0
asyncssh==2.22.0
attrs==26.1.0
autopep8==2.0.4
babel==2.18.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
binaryornot==0.6.0
black==26.3.1
bleach==6.3.0
bokeh==3.9.0
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.0.post2
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cookiecutter==2.7.1
cryptography==46.0.6
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
diff-match-patch==20241021
dill==0.4.1
docstring-to-markdown==0.17
docutils==0.22.4
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
flake8==7.1.2
fonttools==4.62.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.3.0
gmpy2==2.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
imagesize==2.0.0
importlib_metadata==9.0.0
inflection==0.5.1
intervaltree==3.2.1
ipykernel==6.31.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
isort==8.0.1
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jedi==0.19.2
jeepney==0.9.0
jellyfish==1.2.1
Jinja2==3.1.6
joblib==1.5.3
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keyring==25.7.0
kiwisolver==1.5.0
lark==1.3.1
line_profiler==5.0.2
llvmlite==0.47.0
loguru==0.7.3
lsprotocol==2025.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.2.0
more-itertools==11.0.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.19.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.5
notebook_shim==0.2.4
numba==0.65.0
numpy==2.4.4
numpydoc==1.10.0
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
packaging==26.0
pandas==3.0.2
pandocfilters==1.5.1
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.2.0
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycodestyle==2.12.1
pycparser==3.0
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.9.0
Pygments==2.20.0
PyJWT==2.12.1
pylint==4.0.5
pylint-venv==3.0.4
pyls-spyder==0.4.0
PyNaCl==1.6.2
pyparsing==3.3.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.18.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.18
python-dateutil==2.9.0.post0
python-json-logger==4.1.0
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-ruff==2.3.1
python-lsp-server==1.14.0
python-slugify==8.0.4
pytokens==0.4.1
pytoolconfig==1.3.1
pyuca==1.2
pyxdg==0.28
PyYAML==6.0.3
pyzmq==27.1.0
QDarkStyle==3.2.3
qstylizer==0.2.4
QtAwesome==1.4.1
qtconsole==5.7.2
QtPy==2.4.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
roman-numerals==4.1.0
rope==1.14.0
rpds-py==0.30.0
rtree==1.4.1
ruff==0.15.9
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
SecretStorage==3.5.0
Send2Trash==2.1.0
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.8.3
Sphinx==9.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.1.4
spyder-kernels==3.1.4
stack-data==0.6.3
superqt==0.8.1
symengine==0.14.1
sympy==1.14.0
terminado==0.18.1
text-unidecode==1.3
textdistance==4.6.3
threadpoolctl==3.6.0
three-merge==0.1.1
tinycss2==1.4.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.5.0
triton==3.6.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
uri-template==1.3.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
whatthepatch==1.0.7
widgetsnbextension==4.0.15
wurlitzer==3.1.1
xyzservices==2026.3.0
yapf==0.43.0
yarl==1.23.0
zipp==3.23.0
